requirements.txt
pytrends
playwright
google-generativeai
python-dotenv
Pillow
requests
